add redhat_instance service type provider
This provider is a subclass of the core redhat service provider.  It is
of limited use and is intended for service scripts that support managing
multiple service instances via additional arguments to the init script.

This is needed to function with 389's sysvinit script.  It appears that
this type of kluedge will not be nessicary for the systemd service
files.